+++ src/custom_syntax.rs
@@ -65,11 +65,17 @@ impl<'a> From<&'a Expr> for Expression<'a> {
 }
 
 impl Expression<'_> {
-    /// If this expression is a variable name, return it.  Otherwise [`None`].
+    /// Get the value of this expression if it is a variable name or a string constant.
+    ///
+    /// Returns [`None`] also if the constant is not of the specified type.
     #[inline(always)]
     #[must_use]
-    pub fn get_variable_name(&self) -> Option<&str> {
-        self.0.get_variable_name(true)
+    pub fn get_string_value(&self) -> Option<&str> {
+        match self.0 {
+            Expr::Variable(_, _, x) if x.1.is_none() => Some(x.2.as_str()),
+            Expr::StringConstant(x, _) => Some(x.as_str()),
+            _ => None,
+        }
     }
     /// Get the position of this expression.
     #[inline(always)]
@@ -109,6 +115,9 @@ impl Expression<'_> {
         if TypeId::of::<T>() == TypeId::of::<ImmutableString>() {
             return match self.0 {
                 Expr::StringConstant(x, _) => unsafe_try_cast(x.clone()).ok(),
+                Expr::Variable(_, _, x) => {
+                    unsafe_try_cast(Into::<ImmutableString>::into(&x.2)).ok()
+                }
                 _ => None,
             };
         }